% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Activation}
\alias{Activation}
\title{Activation}
\format{A flowSet containing two flowFrame objects:}
\usage{
Activation
}
\description{
A flowSet containg two flowFrames to assess activation of OT-I and OT-II transgenic T cells.
}
\keyword{datasets}
