% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoGate.R
\name{cytoGate}
\alias{cytoGate}
\title{Updated version of DrawGate}
\usage{
cytoGate(fr, channels, alias, subSample = NULL, gate_type = NULL,
  axis = "x", adjust = 1.5, plot = TRUE, labs = TRUE, ...)
}
\arguments{
\item{fr}{a \code{flowFrame} object containing the flow cytometry data for plotting and gating.}

\item{channels}{a vector indicating the fluorescent channel(s) to be used for gating. If a single channel is supplied, a histogram of
of the kernel density will be constructed.}

\item{gate_type}{a character string of length 1 indicating the type of gate to be constructed. Supported gates are \code{"polygon"},
\code{"rectangle"}, \code{"interval"}, \code{"threshold"}, \code{"ellipse"} and \code{"quadrant"}.}

\item{axis}{indicates which axis should be gated for \code{gate_type="interval"} with 2 fluorescent channel supplied.}

\item{adjust}{numeric smoothing factor used for 1D density plots.}

\item{...}{additional arguments for plotDens.}

\item{N}{an integer indicating the number of gates to construct.}
}
\value{
a \code{dataframe} object containing the coordinates required to construct the gate.
}
\description{
\code{DrawGate} implements an interactive manual gating routine for flow cytometry data. Users can easily
select gate coordinates on plots of flow cytometry data using a mouse click. Based on the user input, \code{DrawGate}
can construct many different types of gates, including \code{polygon}, \code{rectangle}, \code{interval}, \code{threshold}, \code{ellipse},
and \code{quadrant}. The type of gate to be constructed must be supplied as the \code{gate_type} argument
which by default is set to a \code{polygonGate}. Each \code{gate_type} has specific gating instructions which are printed to
the console during gating. The selection of multiple gates is supported for rectangle and interval gate types.
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
\keyword{FlowJo,}
\keyword{draw,}
\keyword{gating,}
\keyword{manual,}
\keyword{openCyto}
\keyword{polygonGate,}
