% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawGate.R
\name{gate_draw}
\alias{gate_draw}
\title{DrawGate plugin for openCyto}
\usage{
gate_draw(fr, pp_res, channels, filterId = "", gate_range = NULL,
  min = NULL, max = NULL, gate_type = c("polygon", "rectangle",
  "interval", "threshold", "ellipse", "quadrant"), N = 1, axis = c("x",
  "y"), ...)
}
\arguments{
\item{fr}{a \code{flowFrame} object containing the flow cytometry data for gating.}

\item{pp_res}{output of preprocessing function.}

\item{channels}{a vector of length 2 indicating the channels used to construct the 2D plot.}

\item{filterId}{gate name assigned by openCyto from the \code{gatingTemplate}.}

\item{gate_range}{range in which gate should be constructed (only needed for autogating functions).}

\item{min}{argument passed to \code{truncate_flowFrame} to restrict data to values > \code{min}.}

\item{max}{argument passed to \code{truncate_flowFrame} to restrict data to values < \code{max}.}

\item{gate_type}{type of gate to be constructed, supported types include 
\code{c("polygon", "rectangle", "interval", "threshold", "ellipse", "quadrant")}.}

\item{...}{additional arguments passsed to \code{DrawGate}.}
}
\value{
a \code{polygonGate} constructed from coordinates supplied by \code{DrawGate}.
}
\description{
\code{DrawGate} allows the user to draw polygon gates directly onto plots of flow cytometry data.
Simply left click to gate and right click to close the gate.
}
\examples{
{
library(openCyto)

registerPlugins(fun = gate_draw, methodName = "DrawGate")
listgtMethods()   # check plugin has been registered with openCyto

fs <- read.flowSet(path = "Samples", pattern = ".fcs") # load in .fcs files

gs <- GatingSet(fs) # add flowSet to GatingSet

template <- add_pop(
gs, alias = "Lymphocytes", pop = "+", parent = "root", dims = "FSC-A,SSC-A", gating_method = "DrawGate", gaating_args = "gate_type='ellipse'",
gating_args = "subSample=10000", collapseDataForGating = TRUE, groupBy = 2
)

# gating window will open to construct gate left click vertices on plot and close gate by right click and selecting "stop".

ggcyto(gs[[1]], subset = "root", aes(x = "FSC-A",y = "SSC-A")) + geom_hex(bins = 100) + geom_stats()

}
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
\keyword{gating,}
\keyword{manual,}
\keyword{polygon,}
\keyword{polygonGate}
