% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OTI}
\alias{OTI}
\title{OTI}
\format{A flowSet containing two flowFrame objects:}
\usage{
OTI
}
\description{
A flowSet containg two flowFrames from tryping of OT-I and OT-II transgenic T cells.
}
\keyword{datasets}
