% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuadGate.R
\name{QuadGate}
\alias{QuadGate}
\title{Quadrant Gates for Flow Cytometry Data}
\usage{
QuadGate(fr, channels, ...)
}
\arguments{
\item{fr}{a \code{flowFrame} object containing the flow cytometry data for plotting and gating.}

\item{channels}{a vector of length 2 indicating the fluorescent channels to be used to construct 2D plot
and \code{QuadGates}.}

\item{...}{additional arguments for plotDens.}
}
\value{
a \code{dataframe} object containing the quadrant gate filters.
}
\description{
Simply select a single point designating the center of the 4 rectangleGates. Gate will be constructed in the
following order: bottom left, bottom right, top right and top left.
}
\details{
\code{QuadGate} facilitates manual drawing of quadrant gates onto flow cytometry plots.
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
\keyword{FlowJo,}
\keyword{draw,}
\keyword{gating,}
\keyword{manual,}
\keyword{openCyto}
\keyword{polygonGate,}
