% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawGate.R
\name{DrawGate}
\alias{DrawGate}
\title{Draw 2D Gates on Plots of Flow Cytometry Data}
\usage{
DrawGate(fr, channels, adjust = 1, ...)
}
\arguments{
\item{fr}{a \code{flowFrame} object containing the flow cytometry data for plotting and gating.}

\item{channels}{a vector of length 2 indicating the fluorescent channels to be used to construct 2D plot
and \code{polygonGate}.}

\item{adjust}{numeric passed to \code{density} for 1D gates to adjust smoothing, \code{adjust = 1} by default.}

\item{...}{additional arguments for plotDens.}
}
\value{
a \code{dataframe} object containing the vertices of the polygon gate.
}
\description{
\code{DrawGate} facilitates manual drawing of polygon gates onto flow cytometry plots.
}
\author{
Dillon Hammill (Dillon.Hammill@anu.edu.au)
}
\keyword{FlowJo,}
\keyword{draw,}
\keyword{gating,}
\keyword{manual,}
\keyword{openCyto}
\keyword{polygonGate,}
